Теперь в runSimulationStep передаётся время в секундах

+++ src/BaseMessage.py
@@ -6,12 +6,12 @@ class BaseMessage(metaclass=ABCMeta):
 #
     type: int        # ID класса сообщения
     priority: int    # Приоритет сообщения(Больше -> выше приоритет)
-    time: int        # Время отправления сообщения
+    time: float        # Время отправления сообщения
     sender_ID: int   # ID модели отправителя
     receiver_ID: int # ID модели получателя
 
     @abstractmethod
-    def __init__(self, type: int, priority: int, time: int, sender_ID: int, receiver_ID: int) -> None:
+    def __init__(self, type: int, priority: int, time: float, sender_ID: int, receiver_ID: int) -> None:
         self.type = type
         self.priority = priority
         self.time = time
